Relocate code from tokenize_utils.jl to tokenize.jl
Now that this code is much cleaner and shorter it makes sense to have
the whole lexer in one file.
